python_backend/requirements.txt
flask
python-dotenv
pymongo
pyjwt
twilio
requests
resend
marshmallow
flask-cors
gunicorn
bcrypt
